Deboggage nombreux...
--HG--
extra : convert_revision : svn%3A6a04c5be-4793-4b4c-94f5-c18864c33b64/trunk%4051

+++ cgal_types.h
@@ -35,7 +35,7 @@
 #include "Rectangle_2.h"
 #include "Cercle_2.h"
 
-typedef CGAL::Simple_cartesian<float>              K;
+typedef CGAL::Simple_cartesian<double>              K;
 //typedef CGAL::Exact_predicates_inexact_constructions_kernel K;
 //typedef CGAL::Exact_predicates_exact_constructions_kernel K;
